+++ Stitch/Graph/StitchAI/LLMRecording/LLMRecordingState.swift
@@ -9,15 +9,11 @@ import Foundation
 
 let LLM_COLLECTION_DIRECTORY = "StitchDataCollection"
 
-
-
-
 struct LLMRecordingState: Equatable {
     
     // Are we actively recording redux-actions which we then turn into LLM-actions?
     var isRecording: Bool = false
     
-//    var actions: [LLMAction] = .init()
     var actions: [LLMStepAction] = .init()
     
     var promptState = LLMPromptState()
