+++ Stitch/Graph/StitchAI/StitchAIConstants.swift
@@ -422,205 +422,3 @@ let VISUAL_PROGRAMMING_ACTIONS = """
 }
 
 """
-
-//import Foundation
-//
-//// MARK: - Main Structure
-//struct VisualProgrammingActions: Codable {
-//    let steps: [ProgrammingAction]
-//}
-//
-//// MARK: - Actions
-//enum ProgrammingAction: Codable {
-//    case addNode(AddNodeAction)
-//    case connectNodes(ConnectNodesAction)
-//    case changeNodeType(ChangeNodeTypeAction)
-//    case setInput(SetInputAction)
-//    case addLayerInput(AddLayerInputAction)
-//    
-//    private enum CodingKeys: String, CodingKey {
-//        case stepType = "step_type"
-//    }
-//    
-//    init(from decoder: Decoder) throws {
-//        let container = try decoder.container(keyedBy: CodingKeys.self)
-//        let stepType = try container.decode(String.self, forKey: .stepType)
-//        
-//        switch stepType {
-//        case "add_node":
-//            self = .addNode(try AddNodeAction(from: decoder))
-//        case "connect_nodes":
-//            self = .connectNodes(try ConnectNodesAction(from: decoder))
-//        case "change_node_type":
-//            self = .changeNodeType(try ChangeNodeTypeAction(from: decoder))
-//        case "set_input":
-//            self = .setInput(try SetInputAction(from: decoder))
-//        case "add_layer_input":
-//            self = .addLayerInput(try AddLayerInputAction(from: decoder))
-//        default:
-//            throw DecodingError.dataCorruptedError(forKey: .stepType, in: container, debugDescription: "Unknown step type")
-//        }
-//    }
-//    
-//    func encode(to encoder: Encoder) throws {
-//        var container = encoder.container(keyedBy: CodingKeys.self)
-//        
-//        switch self {
-//        case .addNode(let action):
-//            try container.encode("add_node", forKey: .stepType)
-//            try action.encode(to: encoder)
-//        case .connectNodes(let action):
-//            try container.encode("connect_nodes", forKey: .stepType)
-//            try action.encode(to: encoder)
-//        case .changeNodeType(let action):
-//            try container.encode("change_node_type", forKey: .stepType)
-//            try action.encode(to: encoder)
-//        case .setInput(let action):
-//            try container.encode("set_input", forKey: .stepType)
-//            try action.encode(to: encoder)
-//        case .addLayerInput(let action):
-//            try container.encode("add_layer_input", forKey: .stepType)
-//            try action.encode(to: encoder)
-//        }
-//    }
-//}
-//
-//// MARK: - Action Structures
-//struct AddNodeAction: Codable {
-//    let nodeName: String
-//    let nodeId: UUID
-//    
-//    private enum CodingKeys: String, CodingKey {
-//        case nodeName = "node_name"
-//        case nodeId = "node_id"
-//    }
-//}
-//
-//struct ConnectNodesAction: Codable {
-//    let fromNodeId: UUID
-//    let toNodeId: UUID
-//    let port: Port
-//    
-//    private enum CodingKeys: String, CodingKey {
-//        case fromNodeId = "from_node_id"
-//        case toNodeId = "to_node_id"
-//        case port
-//    }
-//}
-//
-//struct ChangeNodeTypeAction: Codable {
-//    let nodeId: UUID
-//    let nodeType: NodeType
-//    
-//    private enum CodingKeys: String, CodingKey {
-//        case nodeId = "node_id"
-//        case nodeType = "node_type"
-//    }
-//}
-//
-//struct SetInputAction: Codable {
-//    let nodeId: UUID
-//    let value: InputValue
-//    let port: Port
-//    let nodeType: NodeType
-//    
-//    private enum CodingKeys: String, CodingKey {
-//        case nodeId = "node_id"
-//        case value
-//        case port
-//        case nodeType = "node_type"
-//    }
-//}
-//
-//struct AddLayerInputAction: Codable {
-//    let nodeId: UUID
-//    let layerType: LayerType
-//    let port: Port
-//    
-//    private enum CodingKeys: String, CodingKey {
-//        case nodeId = "node_id"
-//        case layerType = "layer_type"
-//        case port
-//    }
-//}
-//
-//// MARK: - Supporting Types
-//enum NodeType: String, Codable {
-//    case number
-//    case text
-//    case boolean
-//}
-//
-//enum LayerType: String, Codable {
-//    case text = "Text"
-//    case image = "Image"
-//    case shape = "Shape"
-//    case color = "Color"
-//}
-//
-//enum LayerPorts: String, Codable {
-//    case text = "Text"
-//    case scale = "Scale"
-//    case shape = "Shape"
-//    case image = "Image"
-//    case position = "Position"
-//    case color = "Color"
-//    case opacity = "Opacity"
-//}
-//
-//enum Port: Codable {
-//    case integer(Int)
-//    case layerPort(LayerPorts)
-//    
-//    init(from decoder: Decoder) throws {
-//        let container = try decoder.singleValueContainer()
-//        if let intValue = try? container.decode(Int.self) {
-//            self = .integer(intValue)
-//        } else if let layerPort = try? container.decode(LayerPorts.self) {
-//            self = .layerPort(layerPort)
-//        } else {
-//            throw DecodingError.dataCorruptedError(in: container, debugDescription: "Port must be either an integer or LayerPorts value")
-//        }
-//    }
-//    
-//    func encode(to encoder: Encoder) throws {
-//        var container = encoder.singleValueContainer()
-//        switch self {
-//        case .integer(let int):
-//            try container.encode(int)
-//        case .layerPort(let layerPort):
-//            try container.encode(layerPort)
-//        }
-//    }
-//}
-//
-//enum InputValue: Codable {
-//    case number(Double)
-//    case string(String)
-//    case boolean(Bool)
-//    
-//    init(from decoder: Decoder) throws {
-//        let container = try decoder.singleValueContainer()
-//        if let doubleValue = try? container.decode(Double.self) {
-//            self = .number(doubleValue)
-//        } else if let stringValue = try? container.decode(String.self) {
-//            self = .string(stringValue)
-//        } else if let boolValue = try? container.decode(Bool.self) {
-//            self = .boolean(boolValue)
-//        } else {
-//            throw DecodingError.dataCorruptedError(in: container, debugDescription: "Value must be a number, string, or boolean")
-//        }
-//    }
-//    
-//    func encode(to encoder: Encoder) throws {
-//        var container = encoder.singleValueContainer()
-//        switch self {
-//        case .number(let double):
-//            try container.encode(double)
-//        case .string(let string):
-//            try container.encode(string)
-//        case .boolean(let bool):
-//            try container.encode(bool)
-//        }
-//    }
-//}
